Refatoração e atualização do projeto para .NET 8.0
Remoção de arquivos desnecessários e adição de novas classes
para implementar padrões de design na criação de produtos e
validações. O código foi reestruturado para melhorar a
organização e legibilidade, utilizando injeção de dependência.
Configurações para Swagger e ajustes nas configurações de
logging foram adicionados. A estrutura do projeto foi aprimorada
com a implementação de um padrão de observador para notificar
eventos de criação de objetos.
