--- api/api.http
@@ -1,6 +1,0 @@
-@api_HostAddress = http://localhost:5180
-
-GET {{api_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
