+++ dotnet-basics.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "src", "src\src.csproj", "{70705332-DFFB-42C0-89FB-FE180F379324}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lib", "src\lib\lib.csproj", "{70705332-DFFB-42C0-89FB-FE180F379324}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src (1)", "src", "{C45E8456-78EE-4FF2-8C0C-5FF7B45D91B6}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{C45E8456-78EE-4FF2-8C0C-5FF7B45D91B6}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "api", "src\api\api.csproj", "{64E4504A-16DD-4198-9FBC-C93FED7F6C31}"
 EndProject
@@ -28,6 +28,7 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{70705332-DFFB-42C0-89FB-FE180F379324} = {C45E8456-78EE-4FF2-8C0C-5FF7B45D91B6}
 		{64E4504A-16DD-4198-9FBC-C93FED7F6C31} = {C45E8456-78EE-4FF2-8C0C-5FF7B45D91B6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
