+++ src/api/Program.cs
@@ -1,0 +1,19 @@
+using api.infra;
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddInfra();
+builder.Services.AddControllers();
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+app.Run();
