+++ src/api/domain/ProductOperations.cs
